+++ yaci-indexer/src/main/java/org/cardanofoundation/rosetta/yaciindexer/YaciIndexerApplication.java
@@ -4,17 +4,16 @@ import com.bloxbean.cardano.yaci.core.config.YaciConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Import;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
 @Import(ConfigurationH2.class)
 @EntityScan({
-    "org.cardanofoundation.rosetta.yaciindexer.model.entity"
+    "org.cardanofoundation.rosetta.yaciindexer.stores.txsize.model"
 })
 @EnableJpaRepositories({
-    "org.cardanofoundation.rosetta.yaciindexer.model.repository"
+    "org.cardanofoundation.rosetta.yaciindexer.stores.txsize.model"
 })
 public class YaciIndexerApplication {
