+++ yaci-indexer/src/main/java/org/cardanofoundation/rosetta/yaciindexer/stores/txsize/model/TransactionSizeEntity.java
@@ -1,4 +1,4 @@
-package org.cardanofoundation.rosetta.yaciindexer.txSizeStore.model;
+package org.cardanofoundation.rosetta.yaciindexer.stores.txsize.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
