+++ yaci-indexer/src/main/java/org/cardanofoundation/rosetta/yaciindexer/stores/txsize/model/TransactionSizeRepository.java
@@ -1,4 +1,4 @@
-package org.cardanofoundation.rosetta.yaciindexer.txSizeStore.model;
+package org.cardanofoundation.rosetta.yaciindexer.stores.txsize.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
